Group the identifier validators by region
